Move BestPolisKMeans into sklearn section of codebase.
